# zooprocess.openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: sebastien.galvagno@imev-mer.fr
  description: API to manage scan from Zooscan
  title: ZooProcess API
  version: v1
servers:
- url: http://zooprocess.imev-mer.fr:8081/v1
...
paths:
...
  /projects/{projectId}/samples/{sampleId}/subsamples/{subSampleId}/process:
    parameters:
    - description: The project id
      example: 655d3062983b92b6e29b3369
      explode: false
      in: path
      name: projectId
      required: true
      schema:
        type: string
      style: simple
    - description: The sample id
      example: 655db3a95ef7f584269ff27b
      explode: false
      in: path
      name: sampleId
      required: true
      schema:
        type: string
      style: simple
    - description: The sub sample id
      example: 655db3a95ef7f584269ff27b
      explode: false
      in: path
      name: subSampleId
      required: true
      schema:
        type: string
      style: simple

#patch:
    #  description: Update the sampleId only observation is authorized

 
    get:
      description: Get the subsample associated to the sample
      operationId: subsamples.process

      responses:
        "200":
          description: Returned the list of subsamples associated to the sample
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
      tags:
        - User
      security:
        - BearerAuth: [ ]
...
components:
  securitySchemes:
    BearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    Process:
      type: object
      # required:
      # properties:
        # id:
          # readOnly: string
...